Shorten wget call
Later script dismisses the time, so I assume that -N, alias for
--timestamping, is here for its secondary purpose, overwriting the file
instead of creating a version with a suffix. File path is known, so -O
can be used instead of -P for directory.

What I'm really after is seeing how much of the system can be linked to
BusyBox, which has a wget applet without -N but with -O. In this rare
case, adding support has the benefit of saving a few characters.