Be more receptive to offline environments
Current behaviour is to update the cached repositories.xml whenever
an operation is performed, regardless of whether we _need_ that
information to continue (e.g. creating a new local repo, adding
a repo by uri, removing an existing repo).

This commit changes this behaviour to only update the cache when
performing operations that use the cache and adds a new `OFFLINE_MODE`
configuration variable that skips fetching repositories.xml entirely.